requirements_dev.txt
# TODO this is duplicated in setup.py, fix that
PyYAML==5.4.1
pytz==2021.1
Jinja2==2.11.3

# DEV
pytest==6.2.2
isort==5.8.0
black==20.8b1
flake8==3.9.0
## The following requirements were added by pip freeze:
appdirs==1.4.4
attrs==20.3.0
click==7.1.2
dataclasses==0.8
importlib-metadata==3.7.3
iniconfig==1.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
regex==2021.3.17
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.1
